move pysar2insarmaps to unavc2oinsarmaps
essentially also remove the old unavco2insarmaps
